// apps/laboratorios/src/app/core/models/estudiantes-oracle.model.ts
export interface EstudiantesOracle {
    periodo: number;
    idEstudiante: string;
    nombre: string;
    numIdProfesor: string;
    codAsignatura: string;
    nomAsignatura: string;
    grupo: number;
    semestre: number;
    fechaInicio: string;
    fechaFin: string;
    codAula: string;
    nomAula: string;
}